move LowestAddressJoinDecider to separate file
